cinnabar: using git-cinnabar in Arcanist
Summary:
Solution proposed in https://admin.phacility.com/PHI572 enhanced with an ability
to work when the diff's base commit is not public (checking for '00..' as a
result of h2git or git2hg).
Checked with `arc diff` and `arc patch` commands.

Test Plan:
It is possible to `arc patch` a diff created from git-cinnabar repository.
One can use the D1211 to test in local "demo" environment.

Reviewers: smacleod, dkl

Reviewed By: smacleod

Bug #: 1459193

Differential Revision: https://phabricator.services.mozilla.com/D1034