cinnabar: Translate git sha to hg for git-cinnabar diffs
Summary:
Solution proposed in https://admin.phacility.com/PHI572 enhanced with an ability
to work when the diff's base commit is not public (checking for '00..' as a
result of h2git or git2hg).
Checked with `arc diff` and `arc patch` commands.

Test Plan: Use arcanist with git-cinnabar repository

Reviewers: smacleod dkl

Subscribers:

Bug #: 1459193